feat: refactor resolve_output_page_idx to page_index module for better modularity
